+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggplot2 (development version)
 
+* `coord_polar()` can have free scales in facets (@teunbrand, #2815).
+
 * The new argument `axes` in `facet_grid()` and `facet_wrap()` controls the
   display of axes at interior panel positions. Additionally, the `axis.labels`
   argument can be used to only draw tick marks or fully labelled axes